+++ src/main/java/com/planbase/pdf/layoutmanager/attributes/TextStyle.kt
@@ -74,6 +74,8 @@ data class TextStyle(val font: PDFont,    // Tf
     /** Average character width (for this font, or maybe guessed) as a positive number in document units */
     val avgCharWidth: Double = avgCharWidth(font, fontSize, characterSpacing)
 
+    val spaceWidth:Double by lazy { stringWidthInDocUnits(" ") }
+
     // Somewhere it says that font units are 1000 times page units, but my tests with
     // PDType1Font.HELVETICA and PDType1Font.HELVETICA_BOLD from size 5-200 show that 960x is
     // pretty darn good.  If we find a font this doesn't work for, we'll have to adjust.